# lib/api-zod/src/generated/types/contract.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { ContractStatus } from "./contractStatus";

export interface Contract {
  id: number;
  clientName: string;
  clientEmail: string;
  contractType: string;
  /** @nullable */
  templateId?: number | null;
  /** @nullable */
  adobeAgreementId?: string | null;
  status: ContractStatus;
  /** @nullable */
  serviceType?: string | null;
  /** @nullable */
  pricingTier?: string | null;
  /** @nullable */
  startDate?: string | null;
  /** @nullable */
  sentAt?: Date | null;
  /** @nullable */
  signedAt?: Date | null;
  /** @nullable */
  expiredAt?: Date | null;
  /** @nullable */
  signedDocumentUrl?: string | null;
  remindersSent: number;
  /** @nullable */
  lastReminderAt?: Date | null;
  /** @nullable */
  contactInquiryId?: number | null;
  createdAt: Date;
  updatedAt: Date;
}